requirements.txt
# tmuxp uses strict requirement versioning so it comes first
tmuxp

celery
click
emport>=1.2.0
gossip
Flask
Flask-Migrate
Flask-SQLAlchemy
Flask-Loopback
Jinja2
gossip
gunicorn
Logbook
psycopg2
PyYAML
pytest
sentinels
virtualenv
